back/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.92.0
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.1
idna==3.6
multidict==6.0.4
openai==0.27.0
pydantic==1.10.13
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.3
